+++ src/main/java/org/remus/simpleoauthserver/controller/AuthorizationEndpoint.java
@@ -40,6 +40,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 import static org.remus.simpleoauthserver.controller.ValueExtractionUtil.extractValue;
+import static org.remus.simpleoauthserver.grants.OAuthGrant.CODE_CHALLENGE;
 
 @Controller
 @RequestMapping(path = "/auth/oauth")
@@ -58,6 +59,7 @@ public class AuthorizationEndpoint {
     private final AuthorizationGrant authFlow;
 
     private final TokenHelper tokenHelper;
+
 
 
     public AuthorizationEndpoint(UserValidator userValidator, LoginAttemptService loginAttemptService, AuthorizationGrant authFlow, TokenHelper tokenHelper) {
@@ -81,6 +83,7 @@ public class AuthorizationEndpoint {
         values.put(STATE, extractValue(requestParams,STATE).orElse(null));
         values.put(CLIENT_ID, extractValue(requestParams, CLIENT_ID).orElseThrow());
         values.put(SCOPE, extractValue(requestParams, SCOPE).orElse(null));
+        values.put(CODE_CHALLENGE, extractValue(requestParams, CODE_CHALLENGE).orElse(null));
         values.put(REDIRECT_URI,  extractValue(requestParams,REDIRECT_URI).orElse(null));
         model.addAttribute("login", new LoginForm(tokenHelper.encode(values)));
         model.addAttribute("appName", application.getName());
@@ -93,11 +96,12 @@ public class AuthorizationEndpoint {
     public String authorizeSubmit(@ModelAttribute("login") @Validated LoginForm login, BindingResult result,
                                   Model model, RedirectAttributes redirectAttributes, HttpServletRequest request) {
         String token = login.getSignedData();
-        Map<String, Object> values = tokenHelper.decode(token, REDIRECT_URI, CLIENT_ID, SCOPE, STATE);
+        Map<String, Object> values = tokenHelper.decode(token, REDIRECT_URI, CLIENT_ID, SCOPE, STATE, CODE_CHALLENGE);
         String redirectUri = (String) values.get(REDIRECT_URI);
         String clientId = (String) values.get(CLIENT_ID);
         String scopeList = (String) values.get(SCOPE);
         String state = (String) values.get(STATE);
+        String codeChallenge = (String) values.get(CODE_CHALLENGE);
         if (loginAttemptService.isBlocked(request.getRemoteAddr())) {
             throw new ResponseStatusException(HttpStatus.FORBIDDEN, "This IP is blocked");
         }
@@ -114,8 +118,10 @@ public class AuthorizationEndpoint {
                     model.addAttribute("registerApp", new AuthorizeApplicationForm(tokenHelper.encode(values)));
                     return "registerapp";
                 } else {
-                    redirectAttributes.addAttribute("code", authFlow.createAuthorizationToken(login.getUserName(),values));
+                    String authorizationToken = authFlow.createAuthorizationToken(login.getUserName(), values);
+                    redirectAttributes.addAttribute("code", authorizationToken);
                     redirectAttributes.addAttribute(STATE, state);
+                    authFlow.checkPkceEntry(codeChallenge,authorizationToken,clientId,redirectUri);
                     return "redirect:" + redirectUri;
                 }
             } catch (UserNotFoundException e) {
@@ -137,6 +143,7 @@ public class AuthorizationEndpoint {
         return "authorize";
     }
 
+
     @PostMapping("/registerApp")
     public String registerApplication(@ModelAttribute("registerApp") AuthorizeApplicationForm form,
                                       Model model, RedirectAttributes redirectAttributes) {
@@ -145,9 +152,12 @@ public class AuthorizationEndpoint {
         String userName = (String) data.get("user");
         String clientId = (String) data.get(CLIENT_ID);
         String state = (String) data.get(STATE);
+        String codeChallenge = (String) data.get(CODE_CHALLENGE);
         authFlow.registerApplication(clientId,userName);
-        redirectAttributes.addAttribute("code", authFlow.createAuthorizationToken(userName,data));
+        String authorizationToken = authFlow.createAuthorizationToken(userName, data);
+        redirectAttributes.addAttribute("code", authorizationToken);
         redirectAttributes.addAttribute(STATE,state);
+        authFlow.checkPkceEntry(codeChallenge,authorizationToken,clientId,redirectUri);
         return "redirect:" + redirectUri;
     }