+++ CHANGELOG.MD
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 ### Added
+* [Support for Authorization with PKCE (rfc7636)](https://github.com/tmseidel/simple-oauth-server/issues/19)
 * [Optimized REST-Calls for Configuration Client #16](https://github.com/tmseidel/simple-oauth-server/issues/16)
 
 ## [0.1.0] - 2021-01-28